+++ test/bitstampIntegrationTests.js
@@ -1,23 +1,23 @@
-var chai = require("chai");
-var sharedTests = require("./shared/integrationTest.js");
+const chai = require("chai");
+const sharedTests = require("./shared/integrationTest.js");
 
 // configure Integration tests variables below this line
 
-var slug = "bitstamp";
-var apiHost = {
+const slug = "bitstamp";
+const apiHost = {
     private: "https://www.bitstamp.net",
     public:  "https://www.bitstamp.net"
 };
 
-var publicMethodsToTest = ["getTicker", "getRate", "getOrderBook"];
-var privateMethodsToTest = ["getFee", "getBalance", "getTransactions"];
-var writeMockResponseFileForMethod = "";
+const publicMethodsToTest = ["getTicker", "getRate", "getOrderBook"];
+const privateMethodsToTest = ["getFee", "getBalance", "getTransactions"];
+const writeMockResponseFileForMethod = "";
 
 // don't change below this line; only configure above this line
 
 
 describe("Integration Test " + slug + ":", function () {
-    var contextIT = {                     // set context for Integration Testing
+    let contextIT = {                     // set context for Integration Testing
         slug: slug,
         apiHost: apiHost,
         publicMethodsToTest: publicMethodsToTest,