+++ test/helpers/config.js
@@ -1,6 +1,6 @@
 var config = {};
 
-config.slug = ["bitfinex", "bitstamp", "bitx", "btce", "newExchangeTemplate", "oxr", "coinbase"];
+config.slug = ["bitfinex", "bitstamp", "bitx", "btce", "newExchangeTemplate", "oxr"];
 
 config.methods = ["getRate", "getTicker", "getOrderBook", "getTrades", "getFee", "getTransactions", "getBalance",
     "getOpenOrders", "postSellOrder", "postBuyOrder", "cancelOrder"];