+++ CHANGELOG.md
@@ -15,6 +15,7 @@ cryptox module adheres to [Semantic Versioning](http://semver.org/).
 - parameter `margin` added to `getOpenOrder` result
 - implement `poloniex.postSellOrder()`
 - implement `poloniex.postBuyOrder()`
+- implement `poloiex.getMarginPositions()`
 
 
 ## [0.2.2] - 2015-11-01